Partially working
Incomplete - I had to ditch the original code (uploading here as processandselect_backup) and just feed channel 1 into the channel 2 code, which is (theoretically) set up for multiple input detection. Added a button to slow down the Trill Bar output because it was making it impossible to read error messages as it just filled the console with bangs, when it's not even attached. This version is now sort-of working, but complaining that all my ch2 arrays are multiply defined, which I suspect may be a problem perhaps of referencing the abstraction that contains the definitions twice - perhaps it's triggering the loadbang each time it's referenced? Might need to go back to that being an inlet (previous error about receiving a set when it wanted a signal was because I was sending line~ and the set to the same outlet (because they go to the same inlet on tabread4~) - separated into an outlet~ for line~ and outlet for set and just combining them in the outer patch has fixed this)