Adjust volume of channel 2
For some reason once the contact mics are attached the signal from channel 2 is coming in as significantly lower than that from channel 1. It's not the mic, and it's not the object it's attached to - seems to be the adaptor (2 x mono 1/4" female to 1 x stereo minijack male). Used this patch to test how much to raise the gain by to get them in the same approximate ballpark - 3x seems to be right.