Testing with 2 contact mics
Ch2 adjusted for volume discrepancy (3* signal) to compensate for the problem with the adapter. However, ch1 taps seem to be triggering sounds from both soundbanks, while ch2 seems to be behaving correctly. I'm wondering if this is a problem with merging the two into a single tabread4~ object - perhaps separating these will make sure that only the correct sounds are being output to the dac~?