All 4 input channels working!
I had missed connecting the detection level to processandselect 3 & 4, also the loadbang to initialise them. Also, in the soundbank initialisation for these channels, I'd missed a send variable name which was channel 2 rather than 3 or 4. With equivalent materials, all mics are now returning their correct sounds although there are discrepancies of level between sounds which are likely to need to be adjusted when we record the proper sounds and there's still the question of probably needing differing detection levels for different objects (higher level possible for hard objects, very low level for soft ones, it looks like) which may need to be hard-coded for reliability and not adjusted in performance, which seems... risky.