midi level control working... BUT
level always starts at zero, which triggers every sound we have in perpetuity. Which sounds AWESOME, but isn't the piece. Feeling at a bit of a loss how to get it to start with a value in the middle of the range, and with the Beatstep's perpetual-twiddle situation, coupled with the total lack of display on the Bela (& while I'm tempted to set up a display for this, TBH I don't think I have time), I can't just turn it on and set it before the performance because I can't see what I'm setting it to. Hmm. Must be a way around this.