MIDI control for response sensitivity
It's working! Note that this does require the EXT SYNC button to be on on the Arturia Beatstep though - without it, it skips back to 0. Numbers seem a little inaccurate - when turning the dial up from "64" it tends to show something like 58, but it's preventing the cacophony of a 0 level to start with and gives a bit more control.