Initial idea for alternate soundbanks
Superseded almost immediately for a (hopefully) more elegant solution which has all the sounds in the one definition file and just changes the parameters of the random number selection to cover a different zone. This one, however, was starting to set up an alternate soundbank as if it were an additional channel, then switch between processandselect modules with MIDI, which would have meant a lot of duplication. Uploaded the sploshing and plasticbag sounds to populate the new soundbank with sounds unlike those already in place.