Disconnected transposition calculation
Trying to fix the pitch shifting where it shouldn't - disconnected the transposition abstraction on channels 2, 3 and 5, but while 5 still seems OK, 2 and 3 are definitely still pitch-shifting, even though there's no logical reason for it. Tried cleaning the project in the Bela IDE, but no change - must be a logic problem somewhere

Too tired to continue tonight. Channels are working, which is the main thing for tomorrow's session. I'll try to look into the logic in a break tomorrow.