Added MIDI switchbank to ch4
Retrieved from main branch, as was selectasample_ch4. Untested with the hardware as yet. Next challenge: How to make the MIDI_switchbank abstraction route to different channels according to the buttons selected. Maybe different outlets & just use the one that applies? Will keep thinking about this. Abstraction doesn't know which subpatch we're in - can it?