individual MIDI levels, 0 level control
Needed to remake file 15 as PD was crashing every time I tried to open it, Moved local repository off OneDrive in case that was causing a problem as local files worked fine. I've created a new abstraction, midi_levels that accepts numeric arguments for MIDI control number and audio channel number so can control individual channel's response. Also added control to allow single-button switching of all channels to 0-response for that chaotic effect that is at the heart of this piece.