Corrupted file replaced, speed improvements
selectasample_ch4 had somehow corrupted & was breaking PD. Replaced with a copy of selectasample_ch3, renamed as ch4 for now - ideally want to go over all these & make a single adaptable abstraction using $ variables for channel number so the patch doesn't need to load 4 separate files but rather only 1 file 4 times - feels like it will be more efficient. Changes made to file 15 (from a backup copy when I couldn't open it with PD because of the corruption issue), removed all occurrences of Graph on Parent for speed improvements (& also because these were mostly safety net checks when I was first developing the patch & don't have a place when it moves over to Bela), tested the MIDI controls, which work OK, but have become stumped by the reset all to zero. The button works to reset the levels to zero - but as soon as you twiddle a knob, it's recalling the value it had before the reset to zero. I can't get ctlin to work to set the knob to also be zero YET.

To-dos: Sort out ctlin to reset knobs to zero; Refine repetitive abstractions to use $ variables to adapt to each channel; Reinstall github application which I may have wrongly blamed for the corruption problem