Improvements & corruption fix
selectasample_ch4.pd had somehow become corrupted, so I've just copied the ch3 version and renamed it for now. Improvements made to 15 for speed, and while the coding changes from earlier this week are working to have MIDI control of the levels for each channel (and the $ variables are working well to allow midi_levels.pd to be adaptable to multiple channels), somehow hitting the button to set all channels to 0 isn't feeding back into the controller, so when you move a pot, it leaps up to whatever value it was on before 0 was set on everything, which is not the behaviour I want.