adaptable version of selectasample
Reconfigured to accept audio channel number as initialising parameter rather than needing a brand new abstraction for every channel. Not yet tested in context, but inital tests with text output seem to suggest it's working. Replaced selectasample variants in all 4 channels of 15. selectasample_flexi.pd and testdrive.pd are the development files for this module. ch 2,3 & 4 versions of selectasample remain untouched - ch1's (which never had a channel number in the title) has been updated to the new version.