Testing flexible version of selectasample
This version seems to work, although I'm feeling it might be dropping the first tap? A bit hard to tell at the moment, but it's triggering sounds, so the basic logic is there and I'm moving on to the next step, which is to test the flexible processandselect. Filename changes for metal bowl and sandpaper sounds because system won't recognise spaces in filenames and inserts underscores, so seemed better to name consistently without spaces to start with, as per the bottle and chain sounds.