Performance improvement attempt
Been experiencing some glitches in playback and lack of responsiveness to handling noise (after no code or setup changes), so have disconnected a number of print statements that seemed to be overloading the Bela console, and made some adjustments - lowered the amplitude threshold for calculating transpositions, and set the lower limit for the initial scale of amplitude which hadn't been set before. Needs to be tested still.